Issue #3:  Support data sources for groups beyond LDAP
