+++ src/main/java/uk/ac/ed/notify/Application.java
@@ -1,25 +1,17 @@
 package uk.ac.ed.notify;
 
-import java.util.HashMap;
-import java.util.Map;
-
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.autoconfigure.velocity.VelocityAutoConfiguration;
 import org.springframework.boot.context.embedded.EmbeddedServletContainerCustomizer;
 import org.springframework.boot.context.embedded.ErrorPage;
-import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.boot.context.web.SpringBootServletInitializer;
 import org.springframework.boot.orm.jpa.EntityScan;
 import org.springframework.cloud.security.oauth2.resource.EnableOAuth2Resource;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.http.HttpStatus;
-import org.springframework.ldap.core.ContextSource;
-import org.springframework.ldap.core.LdapTemplate;
-import org.springframework.ldap.core.support.LdapContextSource;
 
 /**
  * Created by rgood on 18/09/2015.
@@ -33,29 +25,10 @@ import org.springframework.ldap.core.support.LdapContextSource;
 @EnableAutoConfiguration(exclude = VelocityAutoConfiguration.class)
 public class Application extends SpringBootServletInitializer {
 
-    @Value("${java.naming.ldap.derefAliases}")
-    private String derefAliases;
-
     public static void main(String[] args) {
         SpringApplication.run(Application.class, args);
     }
 
-    @Bean
-    @ConfigurationProperties(prefix="ldap.contextSource")
-    public LdapContextSource contextSource() {
-        LdapContextSource contextSource = new LdapContextSource();
-        Map<String,Object> baseEnvironmentProperties = new HashMap<>();
-        baseEnvironmentProperties.put("java.naming.ldap.derefAliases", derefAliases);
-        contextSource.setBaseEnvironmentProperties(baseEnvironmentProperties);
-        return contextSource;
-    }
-
-    @Bean
-    public LdapTemplate ldapTemplate(ContextSource contextSource) {
-        return new LdapTemplate(contextSource);
-    }    
-    
-    
     /* */
     @Bean
     public EmbeddedServletContainerCustomizer containerCustomizer() {
@@ -69,6 +42,5 @@ public class Application extends SpringBootServletInitializer {
             container.addErrorPages(error401Page, error404Page, error500Page);
         };
     }
-   
-    
+
 }
