+++ src/main/java/uk/ac/ed/notify/entity/LdapGroup.java
@@ -27,6 +27,13 @@ public class LdapGroup {
     public void setDescription(String description) {
         this.description = description;
     }
-    
-    
+
+    @Override
+    public String toString() {
+        return "LdapGroup{" +
+                "ou='" + ou + '\'' +
+                ", description='" + description + '\'' +
+                '}';
+    }
+
 }
