+++ src/main/java/uk/ac/ed/notify/groups/OrgUnit.java
@@ -1,0 +1,62 @@
+package uk.ac.ed.notify.groups;
+
+import java.io.Serializable;
+import java.util.ArrayList;
+import java.util.List;
+
+public class OrgUnit implements Serializable {
+
+	private static final long serialVersionUID = 1565656556565L;
+
+	private String orgUnit;
+
+	private List<String> members = new ArrayList<String>();
+
+	public String getOrgUnit() {
+		return orgUnit;
+	}
+
+	public void setOrgUnit(String orgUnit) {
+		this.orgUnit = orgUnit;
+	}
+
+	public List<String> getMembers() {
+		return members;
+	}
+
+	public void setMembers(List<String> members) {
+		this.members = members;
+	}
+
+	@Override
+	public int hashCode() {
+		final int prime = 31;
+		int result = 1;
+		result = prime * result + ((members == null) ? 0 : members.hashCode());
+		result = prime * result + ((orgUnit == null) ? 0 : orgUnit.hashCode());
+		return result;
+	}
+
+	@Override
+	public boolean equals(Object obj) {
+		if (this == obj)
+			return true;
+		if (obj == null)
+			return false;
+		if (getClass() != obj.getClass())
+			return false;
+		OrgUnit other = (OrgUnit) obj;
+		if (members == null) {
+			if (other.members != null)
+				return false;
+		} else if (!members.equals(other.members))
+			return false;
+		if (orgUnit == null) {
+			if (other.orgUnit != null)
+				return false;
+		} else if (!orgUnit.equals(other.orgUnit))
+			return false;
+		return true;
+	}
+
+}
