+++ src/main/java/uk/ac/ed/notify/service/LdapGroupService.java
@@ -4,7 +4,6 @@
  */
 package uk.ac.ed.notify.service;
 
-import java.io.Serializable;
 import java.util.ArrayList;
 import java.util.Enumeration;
 import java.util.List;
@@ -23,27 +22,23 @@ import org.springframework.ldap.core.support.DefaultDirObjectFactory;
 import org.springframework.ldap.core.support.LdapContextSource;
 import org.springframework.ldap.filter.AndFilter;
 import org.springframework.ldap.filter.EqualsFilter;
-import org.springframework.stereotype.Service;
 import uk.ac.ed.notify.entity.LdapGroup;
+import uk.ac.ed.notify.groups.Member;
+import uk.ac.ed.notify.groups.OrgUnit;
+import uk.ac.ed.notify.groups.OrgUnitMembers;
 
-@Service
-public class LdapService {
+public class LdapGroupService implements GroupService {
     
     protected final Log logger = LogFactory.getLog(this.getClass());
     
-    @Value("${ldap.contextSource.url}")
+    @Value("${ldap.contextSource.url:}")
     private String ldapUrl;    
     
     @Autowired
-    LdapTemplate ldapTemplate;
+    private LdapTemplate ldapTemplate;
         
-    /**
-     * Get a group's children group
-     * @param base
-     * @return a list of group
-     */
-    public List<LdapGroup> getNextLevelGroups(String base)
-    {       
+	@Override
+    public List<LdapGroup> getNextLevelGroups(String base) {
         AndFilter andFilter = new AndFilter();
 
         andFilter.and(new EqualsFilter("objectclass","organizationalUnit"));
@@ -67,13 +62,8 @@ public class LdapService {
         }                      
     }
     
-    /**
-     * Get a group's name
-     * @param base
-     * @return name of group
-     */
-    public String getGroupName(String base)
-    {       
+    @Override
+    public String getGroupName(String base) {
         AndFilter andFilter = new AndFilter();
         andFilter.and(new EqualsFilter("objectclass","organizationalUnit"));
 
@@ -95,13 +85,8 @@ public class LdapService {
         }                      
     }
 
-    /**
-     * Get a group's members
-     * @param base
-     * @return a list of member uun
-     */
-    public List<String> getMembers(String base)
-    {       
+    @Override
+    public List<String> getMembers(String base) {
         AndFilter andFilter = new AndFilter();
         andFilter.and(new EqualsFilter("objectclass","groupOfNames"));
         
@@ -130,8 +115,8 @@ public class LdapService {
     
     List<String> results = null;
             
-    public List<String> getMembersFromParentGroup(String base)
-    {               
+	@Override
+    public List<String> getMembersFromParentGroup(String base) {
         /*
          * input  ou=SU685,ou=D685,ou=P5M,ou=ISG3,ou=ISG,ou=UOE,ou=org,ou=grouper2,dc=authorise,dc=ed,dc=ac,dc=uk
          * output          ou=D685,ou=P5M,ou=ISG3,ou=ISG,ou=UOE,ou=org,ou=grouper2,dc=authorise,dc=ed,dc=ac,dc=uk
@@ -162,9 +147,9 @@ public class LdapService {
         
             
         //always clear the array before each invocation 
-        results = new ArrayList<String>();
+        results = new ArrayList<>();
 
-        List<OrgUnitMembers> orgUnitMembers = new ArrayList<OrgUnitMembers>();
+        List<OrgUnitMembers> orgUnitMembers = new ArrayList<>();
 		
         String reqHierarchy = base;
         String terminateToken = ",ou=org,ou=grouper2,dc=authorise,dc=ed,dc=ac,dc=uk";
@@ -211,9 +196,9 @@ public class LdapService {
         }
     }        
 
-    private class MembersAttributeMapper implements AttributesMapper{        
+    private class MembersAttributeMapper implements AttributesMapper {
         @Override
-        public List<String> mapFromAttributes(Attributes attributes) throws NamingException {
+        public List<String> mapFromAttributes(Attributes attributes) {
             List<String> memberof = new ArrayList();
             try{                
                 for (Enumeration vals = attributes.get("member").getAll(); vals.hasMoreElements();) {                    
@@ -221,8 +206,8 @@ public class LdapService {
                     String member = base.substring(base.indexOf("=") + 1, base.indexOf(","));
                     memberof.add(member);
                 }
-            }catch(Exception e){
-                
+            } catch(NamingException e){
+                throw new RuntimeException(e);
             }
             return memberof;
         }
@@ -290,7 +275,7 @@ public class LdapService {
 		for (OrgUnit orgUnit : orgunits) {
 			OrgUnitMembers orgUnitMember = new OrgUnitMembers();
 			orgUnitMember.setOrgunit(orgUnit.getOrgUnit());
-			List<Member> members = new ArrayList<Member>();
+			List<Member> members = new ArrayList<>();
 			if (null != orgUnit.getMembers() && orgUnit.getMembers().size() > 0) {
 				setNameByUUN(orgUnit, members);
 				// adding members to organisation (orgunit)
@@ -325,13 +310,13 @@ public class LdapService {
                  getOrgUnits().add(nextLevelOfGroupMembers); 
                  //level--;
                 
-                 try{
-                 if (amendBase != null && !amendBase.startsWith("ou=SU") &&  getNextLevelGroups(amendBase).size() >0   && level>0) {
-                    traverseAllNodes(amendBase, getNextLevelGroups(amendBase),level); 
-                 }else{
-                     //exit recursive
-                 }   
-                 }catch(Exception e){
+                 try {
+                     if (amendBase != null && !amendBase.startsWith("ou=SU") &&  getNextLevelGroups(amendBase).size() >0   && level>0) {
+                        traverseAllNodes(amendBase, getNextLevelGroups(amendBase),level);
+                     } else {
+                         //exit recursive
+                     }
+                 } catch(Exception e) {
                      logger.error("error traverseAllNodes - " + base + e);
                  }
             }
@@ -352,227 +337,3 @@ public class LdapService {
 
 }
 
-class OrgUnit implements Serializable{
-	
-	private static final long serialVersionUID = 1565656556565L;
-
-	private String orgUnit;
-	
-	private List<String> members = new ArrayList<String>();
-
-	public String getOrgUnit() {
-		return orgUnit;
-	}
-
-	public void setOrgUnit(String orgUnit) {
-		this.orgUnit = orgUnit;
-	}
-
-	public List<String> getMembers() {
-		return members;
-	}
-
-	public void setMembers(List<String> members) {
-		this.members = members;
-	}
-
-	@Override
-	public int hashCode() {
-		final int prime = 31;
-		int result = 1;
-		result = prime * result + ((members == null) ? 0 : members.hashCode());
-		result = prime * result + ((orgUnit == null) ? 0 : orgUnit.hashCode());
-		return result;
-	}
-
-	@Override
-	public boolean equals(Object obj) {
-		if (this == obj)
-			return true;
-		if (obj == null)
-			return false;
-		if (getClass() != obj.getClass())
-			return false;
-		OrgUnit other = (OrgUnit) obj;
-		if (members == null) {
-			if (other.members != null)
-				return false;
-		} else if (!members.equals(other.members))
-			return false;
-		if (orgUnit == null) {
-			if (other.orgUnit != null)
-				return false;
-		} else if (!orgUnit.equals(other.orgUnit))
-			return false;
-		return true;
-	}
-	
-}
-
-class Person implements Serializable{
-	
-	private static final long serialVersionUID = -2723409478365865769L;
-
-	private String fullName;
-	
-	private String surname;
-
-	public String getFullName() {
-		return fullName;
-	}
-
-	public void setFullName(String fullName) {
-		this.fullName = fullName;
-	}
-
-	public String getSurname() {
-		return surname;
-	}
-
-	public void setSurname(String surname) {
-		this.surname = surname;
-	}
-	
-}
-
-class Member implements Serializable {
-	
-	private static final long serialVersionUID = -339684362913276623L;
-
-	private String uun;
-	
-	private String name;
-
-	public String getUun() {
-		return uun;
-	}
-
-	public void setUun(String uun) {
-		this.uun = uun;
-	}
-
-	public String getName() {
-		return name;
-	}
-
-	public void setName(String name) {
-		this.name = name;
-	}
-	
-}
-
-class OrgUnitMembers implements Serializable {
-
-	private static final long serialVersionUID = -3596643333898462959L;
-	
-	private String orgunit;
-	
-	private List<Member> members = new ArrayList<>();
-
-	public String getOrgunit() {
-		return orgunit;
-	}
-
-	public void setOrgunit(String orgunit) {
-		this.orgunit = orgunit;
-	}
-
-	public List<Member> getMembers() {
-		return members;
-	}
-
-	public void setMembers(List<Member> members) {
-		this.members = members;
-	}
-
-}
-
-class OrgHierarchy implements Serializable {
-	
-	private static final long serialVersionUID = 1657657343243L;
-
-	private String code;
-	
-	private String description;
-	
-	private int level;
-
-	public String getCode() {
-		return code;
-	}
-
-	public void setCode(String code) {
-		this.code = code;
-	}
-
-	public String getDescription() {
-		return description;
-	}
-
-	public void setDescription(String description) {
-		this.description = description;
-	}
-
-	public int getLevel() {
-		return level;
-	}
-
-	public void setLevel(int level) {
-		this.level = level;
-	}
-
-	@Override
-	public int hashCode() {
-		final int prime = 31;
-		int result = 1;
-		result = prime * result + ((code == null) ? 0 : code.hashCode());
-		result = prime * result + ((description == null) ? 0 : description.hashCode());
-		result = prime * result + level;
-		return result;
-	}
-
-	@Override
-	public boolean equals(Object obj) {
-		if (this == obj)
-			return true;
-		if (obj == null)
-			return false;
-		if (getClass() != obj.getClass())
-			return false;
-		OrgHierarchy other = (OrgHierarchy) obj;
-		if (code == null) {
-			if (other.code != null)
-				return false;
-		} else if (!code.equals(other.code))
-			return false;
-		if (description == null) {
-			if (other.description != null)
-				return false;
-		} else if (!description.equals(other.description))
-			return false;
-		if (level != other.level)
-			return false;
-		return true;
-	}
-	
-}
-
-class OrgunitUtils {
-
-	public static String buildLdapPathByOrgHierarchy(List<OrgHierarchy> orgunitHierarchy, int size,
-			String ldapContextBasepath) {
-		
-		StringBuilder sb = new StringBuilder();
-		if (size > 0) {
-
-			for (int i = size - 1; i >= 0; i--) {
-				sb.append("ou=");
-				sb.append(orgunitHierarchy.get(i).getCode());
-				sb.append(",");
-			}
-			sb.append(ldapContextBasepath);
-		}
-		return sb.toString();
-	}
-
-}
