+++ src/main/resources/application.properties
@@ -10,11 +10,11 @@ hub.protocol=http
 hub.server=localhost:${server.port}
 #hub.context=
 
-# Notify Database connection settings (HSQLDB example)
-datasource.notify.driver-class-name=org.hsqldb.jdbc.JDBCDriver
-datasource.notify.url=jdbc:hsqldb:hsql://localhost:8887/uPortal
+# Notify Database connection settings (MS SQL SERVER example)
+datasource.notify.driver-class-name=com.microsoft.sqlserver.jdbc.SQLServerDriver
+datasource.notify.url=jdbc:sqlserver://localhost:1433;
 datasource.notify.username=sa
-datasource.notify.password=
+datasource.notify.password=notification-ui@DEV
 
 # Notify Database connection settings (Oracle example)
 #datasource.notify.driver-class-name=oracle.jdbc.OracleDriver
@@ -34,19 +34,20 @@ hibernate.jdbc.batch_size=25
 
 #Liquibase settings
 liquibase.enabled=true
-liquibase.change-log=classpath:/db/changelog/oracle/db.changelog-master.xml
+#liquibase.change-log=classpath:/db/changelog/oracle/db.changelog-master.xml
+liquibase.change-log=classpath:/db/changelog/mssql/db.changelog-master.xml
 liquibase.check-change-log-location=true
-liquibase.default-schema=notify
+#liquibase.default-schema=notify
 liquibase.drop-first=false
 liquibase.url=${datasource.notify.url}
 liquibase.user=${datasource.notify.username}
 liquibase.password=${datasource.notify.password}
 
 # Learn Database connection settings (HSQLDB example)
-datasource.learn.driver-class-name=org.hsqldb.jdbc.JDBCDriver
-datasource.learn.url=jdbc:hsqldb:hsql://localhost:8887/uPortal
-datasource.learn.username=sa
-datasource.learn.password=
+datasource.learn.driver-class-name=${datasource.notify.driver-class-name}
+datasource.learn.url=${datasource.notify.url}
+datasource.learn.username=${datasource.notify.username}
+datasource.learn.password=${datasource.notify.password}
 
 # Learn Database connection settings (Oracle example)
 #datasource.learn.driver-class-name=oracle.jdbc.OracleDriver
@@ -80,8 +81,8 @@ security.ignored=/favicon.ico,/js/**,/ css/**,/webjars /**
 #Zuul proxy settings
 zuul.routes.resource.path=/notification/**
 # CHANGE TO LOCALHOST HERE
-#zuul.routes.resource.url=http://localhost:8085
-zuul.routes.resource.url=https://dev.notify.ws-apps.is.ed.ac.uk:443
+zuul.routes.resource.url=http://localhost:8085
+#zuul.routes.resource.url=https://dev.notify.ws-apps.is.ed.ac.uk:443
 zuul.routes.user.path=/user/**
 proxy.auth.routes.user=oauth2
 proxy.auth.routes.resource=oauth2
@@ -96,10 +97,10 @@ spring.oauth2.resource.tokenInfoUri=https://dev.oauth.ws-apps.is.ed.ac.uk:443/oa
 spring.oauth2.resource.preferTokenInfo= true 
 
 #Log settings logging.file=D:\\log.txt
-logging.level.org.springframework.security=INFO
+#logging.level.org.springframework.security=INFO
 logging.level.uk.ac.ed.notify=INFO
-logging.level.org.hibernate=DEBUG
-logging.level.org.springframework.data.jpa=DEBUG
+#logging.level.org.hibernate=DEBUG
+#logging.level.org.springframework.data.jpa=DEBUG
 #logging.level.org.springframework.web.client.RestTemplate=DEBUG
 
 #Quartz scheduler properties
@@ -132,6 +133,11 @@ ewsDeleteMode=soft
 #soft
 
 
+## LdapGroupService settings
+##
+## Uncomment and provide a value for the 'ldap.contextSource.url' property to enable groups based on
+## LDAP.
+#
 ldap.contextSource.url=ldaps://authorise.is.ed.ac.uk:636
 ldap.contextSource.base=ou=grouper2,dc=authorise,dc=ed,dc=ac,dc=uk
 java.naming.ldap.derefAliases=never
